+++ weblate/machinery/deepl.py
@@ -23,7 +23,8 @@ from django.conf import settings
 from weblate.machinery.base import MachineTranslation, MissingConfiguration
 
 # Weblate as a CAT tool should use v1 API
-DEEPL_API = "https://api.deepl.com/v1/translate"
+DEEPL_TRANSLATE = "https://api.deepl.com/v1/translate"
+DEEPL_LANGUAGES = "https://api.deepl.com/v1/languages"
 
 
 class DeepLTranslation(MachineTranslation):
@@ -33,6 +34,9 @@ class DeepLTranslation(MachineTranslation):
     # This seems to be currently best MT service, so score it a bit
     # better than other ones.
     max_score = 91
+    language_map = {
+        "zh-hans": "zh",
+    }
 
     def __init__(self):
         """Check configuration."""
@@ -40,15 +44,22 @@ class DeepLTranslation(MachineTranslation):
         if settings.MT_DEEPL_KEY is None:
             raise MissingConfiguration("DeepL requires API key")
 
+    def map_language_code(self, code):
+        """Convert language to service specific code."""
+        return super().map_language_code(code.replace("_", "-").upper())
+
     def download_languages(self):
         """List of supported languages is currently hardcoded."""
-        return ("en", "de", "fr", "es", "it", "nl", "pl", "pt", "ru")
+        response = self.request(
+            "post", DEEPL_LANGUAGES, data={"auth_key": settings.MT_DEEPL_KEY},
+        )
+        return [x["language"] for x in response.json()]
 
     def download_translations(self, source, language, text, unit, user, search):
         """Download list of possible translations from a service."""
         response = self.request(
             "post",
-            DEEPL_API,
+            DEEPL_TRANSLATE,
             data={
                 "auth_key": settings.MT_DEEPL_KEY,
                 "text": text,